app/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
motor==3.6.0
python-dotenv==1.0.0
requests==2.31.0
python-dateutil==2.8.2
openai==1.12.0
anthropic==0.39.0
ics==0.7.2
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
pyjwt==2.8.0
email-validator==2.1.0
httpx==0.26.0
google-cloud-firestore==2.14.0
